+++ tests/test_web_api.py
@@ -82,28 +82,6 @@ def test_no_image_params(
     )
 
 
-def test_no_callback_url_jt5(
-    setup_web_client: Tuple[str, str, str, str],
-    web_partner_params: Dict[str, Any],
-    image_params: List[ImageParams],
-    kyc_id_info: Dict[str, str],
-    option_params: OptionsParams,
-) -> None:
-    """callback url is empty and job type is Enhanced KYC"""
-    api_key, partner_id, sid_server, callback_url = setup_web_client
-    web_api = WebApi(partner_id, callback_url, api_key, sid_server)
-    web_partner_params["job_type"] = JobType.ENHANCED_KYC
-    pytest.raises(
-        ValueError,
-        web_api.submit_job,
-        web_partner_params,
-        image_params,
-        kyc_id_info,
-        option_params,
-        True,
-    )
-
-
 def test_no_id_info_params_jt5(
     web_partner_params: Dict[str, Any],
     setup_web_client: Tuple[str, str, str, str],
@@ -128,7 +106,7 @@ def test_no_id_info_params_jt5(
         image_params,
         {},
         option_params,
-        True,
+        False,
     )
 
     web_partner_params["job_type"] = JobType.BIOMETRIC_KYC
@@ -146,28 +124,7 @@ def test_no_id_info_params_jt5(
         image_params,
         {},
         option_params,
-        True,
-    )
-
-
-def test_no_option_params(
-    setup_web_client: Tuple[str, str, str, str],
-    web_partner_params: Dict[str, Any],
-    kyc_id_info: Dict[str, Any],
-    image_params: List[ImageParams],
-) -> None:
-    """Validates when no option params is provided(empty or None)"""
-    api_key, partner_id, sid_server, callback_url = setup_web_client
-    callback_url = ""
-    web_api = WebApi(partner_id, callback_url, api_key, sid_server)
-    pytest.raises(
-        ValueError,
-        web_api.submit_job,
-        web_partner_params,
-        image_params,
-        kyc_id_info,
-        callback_url,
-        True,
+        False,
     )
 
 
@@ -190,29 +147,6 @@ def test_validate_return(
         kyc_id_info,
         option_params,
         False,
-    )
-
-
-def test__validate_options(
-    setup_web_client: Tuple[str, str, str, str],
-    option_params: OptionsParams,
-    web_partner_params: Dict[str, Any],
-    kyc_id_info: Dict[str, str],
-    image_params: List[ImageParams],
-) -> None:
-    """Performs checks for when option_params return_job_status is
-    false"""
-    api_key, partner_id, _, _ = setup_web_client
-    web_api = WebApi(partner_id, "calback_url", api_key, 0)
-    option_params["return_job_status"] = False
-    pytest.raises(
-        ValueError,
-        web_api.submit_job,
-        web_partner_params,
-        image_params,
-        kyc_id_info,
-        option_params,
-        True,
     )
 
 
@@ -263,7 +197,34 @@ def test_success_true_smile_job_type(
         option_params,
         False,
     )
-    #  == {"success": True, "smile_job_id": "0000000857"},)
+
+
+@responses.activate
+def test_show_deprecation_warning_for_use_validation_api(
+    web_partner_params: Dict[str, Any],
+    kyc_id_info: Dict[str, str],
+    image_params: List[ImageParams],
+    option_params: OptionsParams,
+    client_web: WebApi,
+    signature_fixture: Signature,
+) -> None:
+    """check return data for valid smile_job_type when option_params
+    return_job_status is false"""
+    signature = get_signature(signature_fixture)
+    stub_upload_request(signature)
+    signature["timestamp"] = (datetime.now() - timedelta(days=1)).isoformat()
+    stub_get_job_status(signature, True)
+    stub_get_job_status(signature, True, "ERROR MSG")
+    option_params["return_job_status"] = False
+
+    with pytest.deprecated_call():
+        assert client_web.submit_job(
+            web_partner_params,
+            image_params,
+            kyc_id_info,
+            option_params,
+            True,
+        )
 
 
 def test_missing_partner_params_for_at_least_1_param(
@@ -403,7 +364,7 @@ def test_boolean_options_params_non_jt5(
             image_params,
             kyc_id_info,
             option_params,
-            True,
+            False,
         )
     assert str(value_error.value) == "return_history needs to be a boolean"